refactor: consolidate the app's two parallel tab-component families
Settings page was the only consumer of TabGroup (pill/segmented,
ToggleGroup-based) — the rest of the app uses the Tabs-based family
(underline-tabs/responsive-tabs). Migrated it onto ResponsiveTabs and
deleted tab-group.tsx.

Visual change on the Settings page: pill tabs -> underline tabs (with
mobile Select collapse, new for this page). Confirmed with the user
before making it. ui/toggle-group.tsx (the primitive TabGroup was
built on) now has zero app consumers — left in place, it's a shadcn
primitive.
